+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 
 - **Breaking:** `{Global,}Tlsf::INIT`
+- **Breaking:** `Init` (superseded by `ConstDefault` from [`const-default`](https://crates.io/crates/const-default/1.0.0))
 
 ## [0.1.2] - 2021-05-30